+++ packages/native/src/scope.ts
@@ -37,7 +37,7 @@ export class FastEventScope<
 
     }
     prefix: string = ''
-    emitter!: FastEvent<Events>
+    emitter!: FastEvent<any>
     constructor(options?: DeepPartial<FastEventScopeOptions<FinalMeta, Context>>) {
         this._options = Object.assign({}, this._initOptions(options)) as unknown as FastEventScopeOptions<FinalMeta, Context>
     }
@@ -141,8 +141,10 @@ export class FastEventScope<
 
     public emit(type: Types, directive: symbol): void
     public emit(type: string, directive: symbol): void
+    public emit<R = any, T extends Types = Types>(type: T, payload?: Events[T], retain?: boolean): R[]
+    public emit<R = any, T extends string = string>(type: T, payload?: T extends Types ? Events[T] : any, retain?: boolean): R[]
     public emit<R = any>(type: Types, payload?: Events[Types], options?: FastEventListenerArgs<FinalMeta>): R[]
-    public emit<R = any, T extends string = string>(type: T, payload?: T extends Types ? Events[Types] : any, options?: FastEventListenerArgs<FinalMeta>): R[]
+    public emit<R = any, T extends string = string>(type: T, payload?: T extends Types ? Events[T] : any, options?: FastEventListenerArgs<FinalMeta>): R[]
     public emit<R = any>(message: FastEventEmitMessage<Events, FinalMeta>, options?: FastEventListenerArgs<FinalMeta>): R[]
     public emit<R = any, T extends string = string>(message: FastEventEmitMessage<{
         [K in T]: K extends Types ? Events[K] : any
@@ -161,8 +163,11 @@ export class FastEventScope<
         message.type = this._getScopeType(message.type)!
         return this.emitter.emit(message as TypedFastEventMessage<Events, FinalMeta>, options)
     }
-    public async emitAsync<R = any>(type: string, payload?: any, options?: FastEventListenerArgs<FinalMeta>): Promise<[R | Error][]>
-    public async emitAsync<R = any>(type: Types, payload?: Events[Types], options?: FastEventListenerArgs<FinalMeta>): Promise<[R | Error][]>
+    public async emitAsync<R = any, T extends Types = Types>(type: T, payload?: Events[T], retain?: boolean): Promise<[R | Error][]>
+    public async emitAsync<R = any, T extends string = string>(type: T, payload?: T extends Types ? Events[T] : any, retain?: boolean): Promise<[R | Error][]>
+    public async emitAsync<R = any, T extends Types = Types>(type: T, payload?: Events[T], options?: FastEventListenerArgs<FinalMeta>): Promise<[R | Error][]>
+    public async emitAsync<R = any, T extends string = string>(type: T, payload?:  T extends Types ? Events[T] : any, options?: FastEventListenerArgs<FinalMeta>): Promise<[R | Error][]>
+    
     public async emitAsync<R = any>(message: TypedFastEventMessage<Events, FinalMeta>, options?: FastEventListenerArgs<FinalMeta>): Promise<[R | Error][]>
     public async emitAsync<R = any>(): Promise<[R | Error][]> {
         const results = await Promise.allSettled(this.emit.apply(this, arguments as any))